refactor: Moved domain to src/ from src/server
